Setup test environment.
Made changes to prepare the back-end to run tests. Made changes to package.json scripts to run necessary test commands. Made changes to knexfile to make the db a sqlite database. Made some changes to seed files to account for new shape of the database. Wrote tests for Roles router.